+++ frontend/src/component/LandingPage/Carousels/ClientsCarousel/style.css
@@ -1,6 +1,6 @@
-.bg-color {
+/* .bg-color {
     background-color: rgba(255, 197, 57, 0.1);
-}
+} */
 
 /* Important because bootstrap's carousel-inner does not
 set height to 100% itself
