+++ frontend/src/component/LandingPage/Carousels/DirectorsCarousel/style.css
