+++ frontend/src/component/LandingPage/Carousels/TeamsCarousel/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { Carousel } from 'react-responsive-carousel';
-import './teamsCarousel.css';
+import './style.css';
 
 export default function TeamsCarousel() {
     return (
