+++ frontend/src/component/LandingPage/Carousels/TeamsCarousel/style.css
